fixed styles localization
